add batteryinfo as a fragment in batteryratefragment. make some change on cirleprogress to make timer stop at current power(now set 80%)
