+++ ApmPowerManager/app/src/main/java/hult/netlab/pku/apmpowermanager/BatteryRateFragment.java
@@ -1,21 +1,19 @@
 package hult.netlab.pku.apmpowermanager;
 
 
-import android.content.Context;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 
-import android.support.v4.app.FragmentActivity;
+
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentStatePagerAdapter;
+import android.support.v4.view.PagerAdapter;
+import android.support.v4.view.ViewPager;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.BaseAdapter;
-import android.widget.ListView;
-import android.widget.RadioButton;
-import android.widget.TextView;
 
-import java.util.ArrayList;
 import java.util.Timer;
 import java.util.TimerTask;
 
@@ -30,10 +28,12 @@ import java.util.TimerTask;
  */
 public class BatteryRateFragment extends Fragment {
 
+    private static final int NUM_PAGES = 2;
     private Timer timer;
     private DonutProgress donutProgress;
-    private ArrayList<Mode> modeList;
-    private ListView listView;
+    private ViewPager mPager;
+    private PagerAdapter mPagerAdapter;
+
 
     public BatteryRateFragment() {
         // Required empty public constructor
@@ -55,89 +55,54 @@ public class BatteryRateFragment extends Fragment {
         timer = new Timer();
        final MainActivity main = new MainActivity();
         timer.schedule(new TimerTask() {
-            public void run (){
-            main.runOnUiThread(new Runnable() {
+            public void run() {
+                main.runOnUiThread(new Runnable() {
                     @Override
                     public void run() {
                         donutProgress.setProgress(donutProgress.getProgress() + 1);
+                        if (donutProgress.getProgress() == 80) {
+                            timer.cancel();
+                        }
                     }
-                });}
+                });
+            }
 
-        }, 1000, 100);
+        }, 100, 30);
 
-        listView = (ListView)rootview.findViewById(R.id.mode_list);
-        modeList = getData();
-        modeAdapter adapter =new modeAdapter(container.getContext());
-        listView.setAdapter(adapter);
+        mPager = (ViewPager)rootview.findViewById(R.id.sub_pager);
+        mPagerAdapter = new ScreenSlidePagerAdapter(getChildFragmentManager());
+        mPager.setAdapter(mPagerAdapter);
+        mPager.setCurrentItem(1);
+        mPager.setOnPageChangeListener(new ViewPager.SimpleOnPageChangeListener() {
+            @Override
+            public void onPageSelected(int position) {
+
+            }
+        });
+
         return rootview;
     }
 
-    private ArrayList<Mode> getData(){
-        ArrayList<Mode> list = new ArrayList<Mode>();
-        for(int i=1;i<10;i++) {
-            Mode mode = new Mode();
-            mode.check = Math.random()>0.5 ? true : false;
-            mode.name = "mode"+i;
-            mode.description = "mode"+i+" description!";
-            list.add(mode);
-        }
-        return list;
-    }
 
-    public class modeAdapter extends BaseAdapter {
 
-        private LayoutInflater layoutInflater;
-
-        public modeAdapter(Context context){
-            this.layoutInflater=LayoutInflater.from(context);
+    private class ScreenSlidePagerAdapter extends FragmentStatePagerAdapter {
+        public ScreenSlidePagerAdapter(FragmentManager fm) {
+            super(fm);
         }
 
-        @Override
-        public int getCount() {
-            return modeList.size();
-        }
 
-        @Override
-        public Object getItem(int position) {
-            return modeList.get(position);
-        }
-
-        @Override
-        public long getItemId(int position) {
-            return position;
-        }
-
-        @Override
-        public View getView(int position, View convertView, ViewGroup parent) {
-            ViewHloder holder = null;
-            if(convertView == null){
-                holder = new ViewHloder();
-                convertView = layoutInflater.inflate(R.layout.mode_list_detail,null);
-                holder.checkbutton = (RadioButton)convertView.findViewById(R.id.modecheckbutton);
-                holder.name = (TextView)convertView.findViewById(R.id.modename);
-                holder.description = (TextView)convertView.findViewById(R.id.modedescription);
-                convertView.setTag(holder);
-            }else{
-                holder = (ViewHloder)convertView.getTag();
+        public Fragment getItem(int position) {
+            switch(position) {
+                case 0:
+                    return new FragmentBatteryInfo();
+                case 1:
+                    return new FragmentMode();
+                default: return new FragmentMode();
             }
-            holder.checkbutton.setChecked(modeList.get(position).check);
-            holder.name.setText(modeList.get(position).name);
-            holder.description.setText(modeList.get(position).description);
+        }
 
-            return convertView;
+        public int getCount() {
+            return NUM_PAGES;
         }
     }
-
-    class Mode {
-        boolean check;
-        String name;
-        String description;
-    }
-
-    public final class ViewHloder{
-        public RadioButton checkbutton;
-        public TextView name;
-        public TextView description;
-    }
-
 }
