+++ ApmPowerManager/app/src/main/res/layout/fragment_fragment_battery_info.xml
@@ -1,0 +1,19 @@
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="hult.netlab.pku.apmpowermanager.FragmentBatteryInfo">
+
+    <LinearLayout
+        android:id="@+id/battery_info_list_layout"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical">
+
+        <ListView
+            android:id="@+id/battery_info_list"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content">
+        </ListView>
+    </LinearLayout>
+
+</FrameLayout>
