+++ ApmPowerManager/app/src/main/res/layout/mode_list_detail.xml
@@ -11,7 +11,7 @@
         android:layout_marginLeft="20dp"
         android:layout_marginBottom="10dp"
         android:layout_marginTop="10dp"
-        android:shadowColor="#40c4ff"
+        android:shadowColor="#ffffff"
         />
     <LinearLayout
         android:layout_width="wrap_content"
@@ -24,13 +24,15 @@
             android:layout_height="wrap_content"
             android:id="@+id/modename"
             android:text="mode"
-            android:textSize="25sp"/>
+            android:textSize="25sp"
+            android:textColor="#ffffff"/>
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:id="@+id/modedescription"
             android:text="discription"
-            android:textSize="15sp"/>
+            android:textSize="15sp"
+            android:textColor="#ffffff"/>
 
     </LinearLayout>
